% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{code_range_lims}
\alias{code_range_lims}
\title{Translates ranges into lists of (start, end)}
\usage{
code_range_lims(x, listing = NULL)
}
\arguments{
\item{x}{character vector of range values}

\item{listing}{dataframe of NAICS listing to compare against}
}
\description{
Intended for e.g., "123456-8" -> "123456", "123458"
}
\examples{
x <- c("111111", "123456-8", "111-20")
code_range_lims(x)
}
