% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is_naics}
\alias{is_naics}
\alias{assert_naics}
\title{Check that a character vector contains valid NAICS codes}
\usage{
is_naics(naics)

assert_naics(naics)
}
\arguments{
\item{naics}{character vector of candidate codes}
}
\description{
`is_naics` form is a predicate
`assert_naics` form is an assertion
}
