% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codetype.R
\name{naics_code_level}
\alias{naics_code_level}
\title{Determine NAICS code aggregation level (2, 3, 4, 5, or 6-digit)}
\usage{
naics_code_level(naics_code, naics_version = c("2002", "2007", "2012",
  "2017"), naics_listing = naicsmatch::naics_2017)
}
\arguments{
\item{naics_code}{character vector of possibly valid codes}

\item{naics_version}{length-1 character indicating version}

\item{naics_listing}{datrframe with naics code information, defaults to 2017 version

Only one of `naics_version` and `naics_listing` should be supplied. If both are supplied, `naics_listing` is used.}
}
\description{
Determine NAICS code aggregation level (2, 3, 4, 5, or 6-digit)
}
\examples{
select(ex_asm09, naics_2007, naics_label_2007) \%>\% mutate( level = naics_code_level(naics_2007)) \%>\% View()
}
