% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codetype.R
\name{naics_containing}
\alias{naics_containing}
\title{Return containing naics category}
\usage{
naics_containing(naics_code, level = "5-digit")
}
\arguments{
\item{naics_code}{Currently works through substring, only for standard-ish codes}
}
\value{
character vector up to the appropriate level
}
\description{
Return containing naics category
}
