% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listings.R
\name{naics_listing}
\alias{naics_listing}
\title{Retrieve or specify a NAICS listing version}
\usage{
naics_listing(version = NA_character_, listing = NULL)
}
\arguments{
\item{version}{a character string, options are 2017, 2012, 2007, and 2002.}

\item{listing}{a data frame with a `naics` character column or character vector}
}
\value{
data frame version of specified naics listing

Only one of `version` or `listing` should be supplied
}
\description{
Retrieve or specify a NAICS listing version
}
