% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listings.R
\docType{data}
\name{naics_listing_data}
\alias{naics_listing_data}
\alias{naics_2017_listing}
\alias{naics_2012_listing}
\alias{naics_2007_listing}
\alias{naics_2002_listing}
\title{NAICS Code List Datasets}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2196 rows and 2 columns.}
\usage{
naics_2017_listing

naics_2012_listing

naics_2007_listing

naics_2002_listing
}
\description{
This package contains NAICS code list datasets for the NAICS code definitions for 2017, 2012, 2007, and 2002. Codes are included for all levels of aggregation. Code listings are formatted as dataframes with a single `naics` column. They are downloaded from the U.S. Census.
}
\keyword{datasets}
