% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{naics_sankey}
\alias{naics_sankey}
\title{Make a Sankey Diagram of a NAICS Crosswalk}
\usage{
naics_sankey(crosswalk, from = "naics_2007", to = "naics_2012",
  value = "vos09")
}
\description{
Make a Sankey Diagram of a NAICS Crosswalk
}
