% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup-title-descr.R
\name{naics_title}
\alias{naics_title}
\alias{naics_descr}
\title{Lookup Titles and Descriptions of NAICS Codes}
\usage{
naics_title(naics)

naics_descr(naics)
}
\arguments{
\item{naics}{a character vector of NAICS codes to look up}
}
\description{
Convenience function. Focus on most recent NAICS code classifications. Lookup for prior may be added in teh
}
\examples{
naics_title(c("332710", "332722", "332723"))
naics_descr(c("332710", "332722", "332723"))
}
