% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{ranges_to_set_6digit}
\alias{ranges_to_set_6digit}
\title{Determine categories included in NAICS-based ranges}
\usage{
ranges_to_set_6digit(naics_ranges, version = NA_character_,
  listing = NULL)
}
\arguments{
\item{version}{length-1 character indicating version}

\item{listing}{dataframe or character vector with naics code information

Only one of `version` and `listing` should be supplied.

This version specifically deals with ranges (possibly comma delimited) such as are found in BEA IO table data. Trims any *'s.
E.g., "123456-8, 111111" -> "123456", "123457", "123458", "111111"}

\item{naics_code}{character vector of possibly valid codes}
}
\value{
list of character vectors representing the set for each range. For standard codes, same is returned, (for invalid codes, NA is returned??).
}
\description{
Determine categories included in NAICS-based ranges
}
\examples{
naics_ranges <- io_conc_raw_2002$naics_2002_ranges[1:5]
listing <- naicsmatch::naics_2002_listing
ranges_to_set_6digit(
io_conc_raw_2002$naics_2002_ranges[1:5],
listing = naicsmatch::naics_2002_listing
)
}
