% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollups.R
\name{rollup_set_6digit}
\alias{rollup_set_6digit}
\title{Determine categories included in NAICS-based rollups}
\usage{
rollup_set_6digit(naics_code, naics_version = c("2002", "2007", "2012",
  "2017"), naics_listing = naicsmatch::naics_2017)
}
\arguments{
\item{naics_code}{character vector of possibly valid codes}

\item{naics_version}{length-1 character indicating version}

\item{naics_listing}{datrframe with naics code information, defaults to 2017 version

Only one of `naics_version` and `naics_listing` should be supplied. If both are supplied, `naics_listing` is used.

This version specifically deals with 6-digit rollups, such as "12345M"}
}
\value{
list of character vectors representing the set for each rollup. For standard codes, same is returned, (for invalid codes, NA is returned??).
}
\description{
Determine categories included in NAICS-based rollups
}
