% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollups.R
\name{set_to_rollup_6digit}
\alias{set_to_rollup_6digit}
\title{From a list of sets of naics codes, generate rolled up values}
\usage{
set_to_rollup_6digit(naics_set, naics_version = c("2002", "2007", "2012",
  "2017"), naics_listing = naics_2017_listing)
}
\arguments{
\item{naics_set}{list of character vectors}

\item{naics_version}{char}

\item{naics_listing}{tibble}
}
\value{
naics_code
}
\description{
Reverses the operation of `rollup_to_set_6digit`
}
