% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollups.R
\name{rollup_set_6digit}
\alias{rollup_set_6digit}
\title{Determine categories included in NAICS-based rollups}
\usage{
rollup_set_6digit(naics_code, version = NA_character_, listing = NULL)
}
\arguments{
\item{naics_code}{character vector of possibly valid codes}

\item{version}{length-1 character indicating version}

\item{listing}{dataframe or character vector with naics code information

Only one of `naics_version` and `naics_listing` should be supplied.

This version specifically deals with 6-digit rollups, such as "12345M"}
}
\value{
list of character vectors representing the set for each rollup. For standard codes, same is returned, (for invalid codes, NA is returned??).
}
\description{
Determine categories included in NAICS-based rollups
}
